Changes after review and Modal UI fixes
